==> 01_census_data_processing.Rmd <==
---
title: "census_data_processing "
author: "Bouzaher, Julia"
date: "June 29, 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, message=FALSE, warning =FALSE}

#Loading in packages

library(gtools)
library(leaflet)
library(sf)
library(rgeos)
library(stats)
library(ggplot2)
library(tidyr)
library(readr)
library(viridis)
library(readxl)
library(spdplyr)
library(rlist)
library(data.table)
library(fuzzyjoin)
library(dplyr)
library(readr)
library(data.table)
library(stringr)
library(purrr)

#adjusting memory limit to make R work faster
memory.limit(2000000000)

```


```{r, message=FALSE, warning = FALSE}

#Creating the file path lists necessary to load in all the tables by type and year in a loop
base <- ("E:/Sandusky/data/source/external/DP tables [tract] [county]")
years.list <- list.files("E:/Sandusky/data/source/external/DP tables [tract] [county]")
tabletype.list <- list.files("E:/Sandusky/data/source/external/DP tables [tract] [county]/2010")
list3 <- list.files("E:/Sandusky/data/source/external/DP tables [tract] [county]/2010/DP02 social")

```

This is the main loop that loads all of the Census data separately - every year, every type, and every metadata file that matches variable ID to variable description. The output is 32 individual tables by year and type with labels (ex. df_2010_DP02) and their respecive metadata tables (df_2010_DP02_meta).

One section of the loop has been commented out; that section replaces the coded column names that tell you nothing about the variable (ex. DP02_0111E) with their matching descriptions that tell you what the variable is (ex. Household Income for Single-Parent Households). I commented it out because the descriptive names themselves are still too long and complicated, so on my first pass at constructing my tables I manually handpicked and renamed variables in a new dataframe, but it may be useful to revisit later if I want to clean up all of the Census variables to make the variable names highly readable.


```{r, message=FALSE, warning = FALSE}

#Initializing a dataframe to store sums of NA values
  sums <- data.frame()

for (i in years.list){
  for (j in tabletype.list){
      folderpath <- file.path(base,i,j)
      table.extension <- grep("overlays", list.files(folderpath), value=TRUE)
      metadata.extension <- grep("metadata", list.files(folderpath), value=TRUE)
      path_metadata <- file.path(base,i,j,metadata.extension)
      path <- file.path(base,i,j,table.extension)
      df <- (readr::read_csv(path))
      df_metadata <- (readr::read_csv(path_metadata))
      df <- mutate(df, year = i)
      b <- stringr::str_split(j, "")
      c <- unlist(b)
      d <- paste(c[1:4],collapse="")
      df <- assign(paste0("df_", i, "_",d), df)
      df_metadata <- assign(paste0("df_", i, "_",d,"_metadata"), df_metadata)
  }  
}

```

Next, now that we have the 32 individual tables, I am going to write a second loop that will go through all of the annual versions of a given table type (ex. DP02 - Social Characteristics table), select and rename the given variables that I determine essential to the final aggregate table, and then bind all of the annual table versions together into one long time series data table for that given table type so that there are four time series table types with all years: DP02, DP03, DP04, DP05. Then my final step will be to combine all of those together for my final table. 

I am doing this so that, in addition to having an "inital_table" identifier column, it is easier to keep track of which variables in the final table came from which original table and year in case any individual or intermediate table has to be modified or cleaned since not all variables are constant from year to year.

```{r}
#Attempting to join all the files together
DP02 <- data.frame()
bd = data.frame()
DP_years = 0:8

#Variables needed from DP02: education, sex

#loop through all years
for (i in DP_years){

  #call each year and its metadata and store it in the loop
  bd <- get(paste0("df_201",i,"_DP02"))
  bd_metadata <- get(paste0("df_201",i,"_DP02_metadata"))
  
  #get all education variables
  bd2 <- bd %>% 
   dplyr::select(which(str_detect(bd[1,], "^Estimate!!EDUCATIONAL ATTAINMENT")))
  
  #rename education variables
  bd2names <- (bd2[1,]) 
  bd2 <- bd2[-1,]  
  
  #get all other variables
  bd3 <- bd %>%
    mutate(initial_table_id = paste0("df_201",i,"_DP02"), id = GEO_ID, tract = NAME, year = year) %>%
    dplyr::select(initial_table_id, id, tract, year)
  
  bd3 <- bd3[-1,]
  
  names(bd2) <- bd2names
  bd4 <- cbind(bd3,bd2)
  
  DP02 <- smartbind(bd4, DP02)
  
}

names(DP02) <- sub("^Estimate!!EDUCATIONAL ATTAINMENT!!", "", names(DP02))
  
```

With the DP02 table constructed, we can construct similar tables with the same automatic technique for the DP03 data.


```{r}
#Variables needed from DP03: income, occupation, industry, employment status, poverty level, mean income, median income, income blocks

DP03 <- data.frame()
bd = data.frame()
DP_years = 0:8

#loop through all years
for (i in DP_years){

  #call each year and its metadata and store it in the loop
  bd <- get(paste0("df_201",i,"_DP03"))
  bd_metadata <- get(paste0("df_201",i,"_DP03_metadata"))

  #get all education variables
  bd2 <- bd %>% 
   dplyr::select(which(str_detect(bd[1,], "^Estimate!!INCOME.+Mean household income|^Estimate!!INCOME.+Median household income|^Percent.+SNAP|^Percent!!OCCUPATION|^Percent!!EMPLOYMENT STATUS!!Not|
                           
                           Estimate!!PERCENTAGE.+POVERTY")))
  
  #rename education variables
  bd2names <- (bd2[1,]) 
  bd2 <- bd2[-1,]  
  
  #get all other variables
  bd3 <- bd %>%
    mutate(initial_table_id = paste0("df_201",i,"_DP03"), id = GEO_ID, tract = NAME, year = year) %>%
    dplyr::select(initial_table_id, id, tract, year)
  
  bd3 <- bd3[-1,]
  
  names(bd2) <- bd2names
  bd4 <- cbind(bd3,bd2)
  
  DP03 <- smartbind(bd4, DP03)
  
}


```


With the DP02 table constructed, we can construct similar tables with the same automatic technique for the DP04 data.


```{r}
#Variables needed from DP04: total housing units, year moved into unit, rent, occupied housing units

DP04 <- data.frame()
bd = data.frame()
DP_years = 0:8

#loop through all years
for (i in DP_years){

  #call each year and its metadata and store it in the loop
  bd <- get(paste0("df_201",i,"_DP04"))
  bd_metadata <- get(paste0("df_201",i,"_DP04_metadata"))

  #get all education variables
  bd2 <- bd %>% 
   dplyr::select(which(str_detect(bd[1,], "^Estimate!!HOUSING OCCUPANCY!!Total housing units!!Occupied housing units|^Estimate!!UNITS IN STRUCTURE!!Total housing units|^Estimate!!YEAR STRUCTURE BUILT|^Estimate!!YEAR HOUSEHOLDER MOVED INTO UNIT!!|^Estimate!!GROSS RENT!!")))
  
  #rename education variables
  bd2names <- (bd2[1,]) 
  bd2 <- bd2[-1,]  
  
  #get all other variables
  bd3 <- bd %>%
    mutate(initial_table_id = paste0("df_201",i,"_DP04"), id = GEO_ID, tract = NAME, year = year) %>%
    dplyr::select(initial_table_id, id, tract, year)
  
  bd3 <- bd3[-1,]
  
  names(bd2) <- bd2names
  bd4 <- cbind(bd3,bd2)
  
  DP04 <- smartbind(bd4, DP04)
  
 
}

```


With the DP02 table constructed, we can construct similar tables with the same automatic technique for the DP05 data.


```{r}

#Variables needed from DP05: race, ancestry
DP05 <- data.frame()
bd = data.frame()
DP_years = 0:8

#loop through all years
for (i in DP_years){

  #call each year and its metadata and store it in the loop
  bd <- get(paste0("df_201",i,"_DP05"))
  bd_metadata <- get(paste0("df_201",i,"_DP05_metadata"))

  #get all education variables
  bd2 <- bd %>% 
   dplyr::select(which(str_detect(bd[1,], "^Estimate!!RACE|^Estimate!!ANCESTRY|^Estimate!!SEX AND AGE")),
                 which(str_detect(bd[1,], "^Total Population*")))
  
  #rename education variables
  bd2names <- (bd2[1,]) 
  bd2 <- bd2[-1,]  
  
  #get all other variables
  bd3 <- bd %>%
    mutate(initial_table_id = paste0("df_201",i,"_DP05"), id = GEO_ID, tract = NAME, year = year) %>%
    dplyr::select(initial_table_id, id, tract, year)
  
  bd3 <- bd3[-1,]
  
  names(bd2) <- bd2names
  bd4 <- cbind(bd3,bd2)
  
  DP05 <- smartbind(bd4, DP05)
  
}

```

Now that all of the aggregated time series tables have been constructed by type, we can remove all of the earlier tables.

```{r}
rm(list= ls()[! (ls() %in% c("DP02","DP03","DP04","DP05"))])

```

Defining the variables which we will use to merge our tables as "year" and "tract" (if we just picked year, our merge might muddle all the tracts together and vice versa), we merge all table types together, make sure year is treated as a numeric variable, and then we select only the variables we want. We use str_detect() to select all variables named with a certain topic (since there are different varieties for percentages/estimates, and then use which() to extract the ID, not the boolean, for dplyr::select() to operate on.

```{r, error=FALSE, warning=FALSE}
mergecols <- c("year", "tract")

final <- list(DP02,DP03,DP04,DP05) %>% 
  reduce(full_join, by = mergecols) %>%
  mutate(year = as.numeric(year))

final3 <- final %>%
  dplyr::select(year, 
                tract, 
                which(str_detect(names(final), "Estimate!!YEAR STRUCTURE BUILT!!Total housing units!!Built 2014 or later")),
                which(str_detect(names(final), "Service occupations")),
                which(str_detect(names(final), "Sales and office occupations")),
                which(str_detect(names(final), "Management, business, science, and arts occupations")),
                which(str_detect(names(final), "Percent!!OCCUPATION!!Natural resources, construction, and maintenance occupations")),
                which(str_detect(names(final), "Production, transportation, and material moving occupations")),
                which(str_detect(names(final), "Estimate!!YEAR HOUSEHOLDER MOVED INTO UNIT!!Occupied housing units")),
                which(str_detect(names(final), "Total housing units")))

final2 <- final %>%
  dplyr::select(year, tract, which(str_detect(names(final), "RACE")))
                
final <- final %>%
  dplyr::select(year, tract, which(str_detect(names(final), "Black or African American")), which(str_detect(names(final), "GROSS RENT*.*Median")), which(str_detect(names(final), "Moved in 2015 or later")), which(str_detect(names(final), "Bachelor's degree")), which(str_detect(names(final), "bachelor's")), which(str_detect(names(final), "Median household income")), which(str_detect(names(final), "Mean household income")),  which(str_detect(names(final), "SNAP benefits")),  which(str_detect(names(final), "Moved in 2017 or later")),  which(str_detect(names(final), "Built 2010 or later")), `Estimate!!RACE!!Total population`)

#all variables besides year and tract
vars <- names(final[,c(3:36)])

```

Now, we convert all non-basic (year,tract) variables to numeric values, remove NAs, and merge identical columns that received different names over various iterations of Census data. Lastly, select all the variables needed.

```{r, error=FALSE, warning=FALSE}

final <- final %>%  
  mutate_at(vars, as.numeric)

final <- final %>%
  mutate(MedianRent = rowSums(final[vars[c(9,10)]], na.rm = TRUE)) %>%
  mutate("Moved in 2015 or later" = rowSums(final[vars[11]], na.rm = TRUE)) %>%
  mutate("Moved in 2017 or later" = rowSums(final[vars[31]], na.rm = TRUE)) %>%
  mutate("Built 2010 or later" = rowSums(final[vars[c(32:33)]], na.rm = TRUE)) %>%
  mutate("SNAP benefits" = rowSums(final[vars[c(24:30)]], na.rm = TRUE)) %>%
  mutate("Bachelor's degree" = rowSums(final[vars[c(12:16)]], na.rm = TRUE)) %>%
  mutate("Median household income" = rowSums(final[vars[c(17:23)]], na.rm = TRUE)) %>%
  mutate("Mean household income" = rowSums(final[vars[c(26:34)]], na.rm = TRUE)) %>%
  mutate(year = final$year) %>%
  mutate(NAMELSAD = sub("*, Erie.*", "", final$tract))

```


Now provide more estimates for race/ancestry for all groups.

```{r, error=FALSE, warning=FALSE}

vars2 <- names(final2 %>% dplyr::select(-c(year,tract)))
codes <- read_excel("data/derived/Race_Coding.xlsx") %>%
  mutate(Variable = gsub("\"", "", Variable)) %>%
  mutate(Variable = gsub("\\[|\\]", "", Variable)) %>%
  mutate(Variable = gsub("[0-9]*", "", Variable)) %>%
  mutate(Variable = substr(Variable, 2, nchar(Variable)))

codes <- codes %>%
  
  mutate(Black = ifelse((codes$`race 1` == "Black" | codes$`race 2` == "Black" | codes$`race 3` == "Black"),
                        1, 0),
         
         White = ifelse((codes$`race 1` == "White" | codes$`race 2` == "White" | codes$`race 3` == "White"),
                        1, 0),
         
         Asian = ifelse((codes$`race 1` == "Asian" | codes$`race 2` == "Asian" | codes$`race 3` == "Asian"),
                        1, 0),
         
         Native_American = ifelse((codes$`race 1` == "Native American" | codes$`race 2` == "Native American" | codes$`race 3` == "Native American"),
                        1, 0),
         
         Native_Hawaiian_or_Pacific_Islander = ifelse((codes$`race 1` == "Native Hawaiian or Pacific Islander" | codes$`race 2` == "Native Hawaiian or Pacific Islander" | codes$`race 3` == "Native Hawaiian or Pacific Islander"),
                        1, 0),
         
         Other_Race_Ethnicity = ifelse((codes$`race 1` == "Other Race/Ethnicity" | codes$`race 2` == "Other Race/Ethnicity" | codes$`race 3` == "Other Race/Ethnicity"),
                        1, 0))
  
Black <- codes$Variable[codes$Black==1]
Native_American <- codes$Variable[codes$Native_American==1]
White <- codes$Variable[codes$White==1]
Other <- codes$Variable[codes$Other_Race_Ethnicity==1]
Asian <- codes$Variable[codes$Asian==1]
Native_Hawaiian_or_Pacific_Islander <- codes$Variable[codes$Native_Hawaiian_or_Pacific_Islander==1]

final2 <- final2 %>%  
  mutate_at(vars2, as.numeric)

final2 <- final2 %>%
  mutate(Black = rowSums(final2[,Black], na.rm = TRUE)) %>%
  mutate(White = rowSums(final2[,White], na.rm = TRUE)) %>%
  mutate(Native_Hawaiian_or_Pacific_Islander = rowSums(final2[,Native_Hawaiian_or_Pacific_Islander], na.rm = TRUE)) %>%
  mutate(Asian = rowSums(final2[,Asian], na.rm = TRUE)) %>%
  mutate(Native_American = rowSums(final2[,Native_American], na.rm = TRUE)) %>%
  mutate(Other = rowSums(final2[,Other], na.rm = TRUE)) %>%
  mutate(year = final2$year) %>%
  mutate(Total_Pop = `Estimate!!RACE!!Total population`)

final2 <- final2 %>%
  dplyr::select(year, tract, Black, White,Native_Hawaiian_or_Pacific_Islander,Asian,Native_American,Other, Total_Pop)

```


Lastly, let us extract any other interesting variables we've missed so far.

```{r}
vars3 <- names(final3 %>% dplyr::select(-c(year,tract)))

final3 <- final3 %>%  
  mutate_at(vars3, as.numeric)

final3 <- final3 %>%
  #mutate(MedianRent = rowSums(final[vars[c(9,10)]], na.rm = TRUE)) %>%
  mutate(Total_Housing_Units = rowSums(final3[vars3[c(2,12)]], na.rm = TRUE)) %>%
  mutate(Built_2014_or_Later = rowSums(final3[vars3[1]], na.rm = TRUE)) %>%
  mutate(Built_2010_or_Later = rowSums(final3[vars3[c(36,48)]], na.rm = TRUE)) %>%
  mutate(Moved_In_2010_or_Later = rowSums(final3[vars3[c(18,12:14)]], na.rm = TRUE)) %>%
  mutate(Percent_Service_Occupations = rowSums(final3[vars3[c(2,3)]], na.rm = TRUE)) %>%
  mutate(Percent_Sales_and_Office_Occupations = rowSums(final3[vars3[c(4,5)]], na.rm = TRUE)) %>%
  mutate(Percent_Management_Business_Science_Arts_Occupations = rowSums(final3[vars3[c(6,7)]], na.rm = TRUE)) %>%
  mutate(Percent_Production_MaterialMoving_Transportation_Construction_Occupations = rowSums(final3[vars3[c(8,9,10)]], na.rm = TRUE)) %>%
  dplyr::select(year,tract, Total_Housing_Units, Built_2014_or_Later, Built_2010_or_Later, Moved_In_2010_or_Later, Percent_Service_Occupations, Percent_Production_MaterialMoving_Transportation_Construction_Occupations, Percent_Management_Business_Science_Arts_Occupations,Percent_Sales_and_Office_Occupations)

```


For variables that are all 0 for a given year (group by year), replace the 0s with all NAs. Otherwise, the Shiny App will display zeros and indicate a drop in value instead of displaying that the data is unavailable for a given year.

```{r}
final <- full_join(final, final2, by = c("year", "tract"))

final <- full_join(final, final3, by = c("year", "tract"))

final <- final %>% group_by(year) %>% mutate(dplyr::across(where(is.numeric), ~ replace(., sum(.) == 0, NA))) %>% ungroup() %>%
  dplyr::select(year, tract, Total_Pop, Black, White, Native_Hawaiian_or_Pacific_Islander, Asian, Native_American, Other, "MedianRent","Moved in 2015 or later","Moved in 2017 or later","Built 2010 or later","SNAP benefits","Bachelor's degree","Median household income", "Mean household income", Total_Housing_Units, Built_2014_or_Later, Built_2010_or_Later, Moved_In_2010_or_Later, Percent_Service_Occupations, Percent_Production_MaterialMoving_Transportation_Construction_Occupations, Percent_Management_Business_Science_Arts_Occupations,Percent_Sales_and_Office_Occupations, NAMELSAD)

fwrite(final, file="data/derived/final_01.txt")

rm(list=ls())
```
