==> 02_spatial_data_processing.Rmd <==
---
title: "spatial_data_processing"
author: "Bouzaher, Julia"
date: "September 22, 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Let's test the shapefile.

```{r}
library(rgeos)
library(sp)
library(sf)
library(rgeos)
library(magrittr)
library(dplyr)
library(readr)
library(gridExtra)
library(ggplot2)
library(data.table)
library(rgdal)
library(lubridate)
library(stringr)
library(rgdal)
library(sf) 
library(sp)
library(rgeos)
library(leaflet)
library(raster)
library(dplyr)
library(lwgeom)
library(spatialEco)
library(stats)
library(viridis)
```

Load in shapefile and filter for Erie County. Check the coordinate reference systems. Join the shapefile with the "final" census dataset by tract and remove unnecessary shapefile features. Then, reset coordinate reference system to 4326 again.

```{r}
final <- fread(file="data/derived/final_01.txt")

shape <- read_sf("data/source/external/tract shapefiles [tract]/tl_2019_39_tract.shp") %>%
  filter(COUNTYFP =="043")

shape <- st_transform(shape, 4326)

crs(shape)
crs(final)

factpal <- colorFactor(palette = magma(19), domain = shape$TRACTCE)


final <- full_join(final,shape, by="NAMELSAD") %>%
#  mutate(tract = NAMELSAD) %>%
  dplyr::select(-(c(STATEFP,COUNTYFP,TRACTCE,GEOID,NAME,NAMELSAD,MTFCC,FUNCSTAT,INTPTLON,INTPTLAT)))

final <- st_sf(final)
final <- st_transform(final, 4326)

```

Construction of a sample map. Map 1: Distribution of Black/African American Population

```{r}

factpal <- viridis(length(final$Black))
factpal <- colorNumeric(palette = viridis(997), c(0,997))

#final <- final %>% filter(year == "2016")

leaflet(final) %>%
  addProviderTiles(providers$OpenStreetMap) %>%
  addPolygons(weight=1.4, fillColor = ~factpal(as.numeric(final$Black) ), fillOpacity=1) %>%
  addLegend(position ="bottomleft", pal = factpal, values = as.numeric(final$Black), title = "Number of Black Individuals",labFormat = labelFormat(suffix = " individuals"), opacity=1)


```

Construction of a sample map. Map 2: Distribution of Median Rents

```{r}

range(final$MedianRent)
factpal <- colorNumeric(palette = viridis(500), c(500,1000))

map <- leaflet(final %>% filter(year == "2018")) %>%
  addTiles() %>%
  addProviderTiles(providers$OpenStreetMap) %>%
  addPolygons(label = ~paste ("Tract:",tract, "\nMedian Rent:",MedianRent),
              fillColor = ~factpal(as.numeric(MedianRent)),
              weight = 1,
              smoothFactor = 0.5,
              opacity = 1.0,
              fillOpacity = .88,
              highlightOptions = highlightOptions(color = "white",
                                                  weight = 2,
                                                  bringToFront = TRUE)) %>%
  addLegend(position ="bottomleft", pal = factpal, values = as.numeric(final$MedianRent), title = "Median Rents 2018",labFormat = labelFormat(prefix = "$"), opacity=1)

```

Construction of a sample map. Map 3: View the GLRI Pathway Polygons.

```{r}
GLRI <- read_sf("data/source/city/GLRI Pathway/Polygon_layer.shp")
GLRI <- st_transform(GLRI, 4326)

leaflet(GLRI) %>%
  addProviderTiles(providers$OpenStreetMap) %>%
  addPolygons()

```

Construction of a sample map. Map 4: View the Neighborhood Polygons.

```{r}
Neighborhoods <- read_sf("data/source/city/Neighborhoods/City_of_Sandusky_Neighborhoods.shp")
Neighborhoods <- st_transform(Neighborhoods, 4326)

leaflet(Neighborhoods) %>%
  addProviderTiles(providers$OpenStreetMap) %>%
  addPolygons()

```

Load in Wetlands data, Parks data, and Leaf Icons. Map 5: View the Wetlands data, Parks data, and Leaf Icons.

```{r}

#wetlands data as a result of two Wetlands areas being merged
Wetlands <- read_sf("data/source/external/wetlands data/HU8_04100011_Watershed/HU8_04100011_Wetlands.shp")
Wetlands2 <- read_sf("data/source/external/wetlands data/HU8_04100012_Watershed/HU8_04100012_Wetlands.shp")
Wetlands <- st_transform(Wetlands, CRS("+init=epsg:4326"))
Wetlands2 <- st_transform(Wetlands2, CRS("+init=epsg:4326"))
Wetlands3 <- rbind(Wetlands, Wetlands2)
Wetlands3 <- st_crop(Wetlands3, xmin = -82.918968, xmax = -82.314877,  ymin = 41.277963, ymax = 41.483993)

#parks names data
parks <- fread("data/source/external/Parks/OH_Features_20200901.txt") %>%
  filter(COUNTY_NAME == "Erie") %>%
  filter(FEATURE_CLASS == "Park") %>%
  rename(lat = PRIM_LAT_DEC, long = PRIM_LONG_DEC)

#land cover data
leafIcons <- icons(
  iconUrl = 
    "http://leafletjs.com/examples/custom-icons/leaf-green.png",
  iconWidth = 15, iconHeight = 40,
  iconAnchorX = 14, iconAnchorY = 40
)

leaflet() %>%
  addProviderTiles(providers$OpenStreetMap) %>%
addPolygons(data = Wetlands3, group = "Wetlands", stroke = TRUE, weight=.4, opacity = 1) %>%
addMarkers(data = parks, icon = leafIcons, group = "Parks", popup = ~(FEATURE_NAME))


```

Now, let's begin the raster processing. To read in all of the rasters:

```{r}

landcover2016 <- raster::raster("data/source/external/landcover data/US_200EVC/us_200evc/w001001.adf")

landcover2014 <- raster::raster("data/source/external/landcover data/US_140EVC/us_140evc/w001001.adf")

landcover2012 <- raster::raster("data/source/external/landcover data/US_130EVC/us_130evc/w001001.adf")

landcover2010 <- raster::raster("data/source/external/landcover data/US_120EVC/us_120evc/w001001.adf")

landcover2008<- raster::raster("data/source/external/landcover data/US_110EVC/us_110evc/w001001.adf")

landcover2001 <- raster::raster("data/source/external/landcover data/US_105EVC/us_105evc/w001001.adf") 

```


Filter/Mask for non-desired categories for each of the raster years. 
A common quartile-esque labels system was determined for consistency across years (some years had 10% intervals, other years did not) and
rasters were reclassified accordingly. Note the use of c() to construct a proper reclassification matrix.

```{r}
#landcover 2001
landcover2001@data@attributes[[1]]$ID
landcover2001@data@attributes[[1]]$CLASSNAMES
length(landcover2001@data@attributes[[1]]$ID)

a <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")]
b <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")] 
d <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]   
f <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
g <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
h <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Orchard")]
i <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop-Close Grown Crop")]
j <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
k <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Close Grown Crop")]
l <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Fallow/Idle Cropland")]
m <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "NASS-Pasture and Hayland")]
n <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Pasture/Hay")]
o <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Cultivated Crops")]
p <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 10 and < 20%")]
q <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 40 and < 50%")]
r <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 90 and <= 100%")]
s <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 10 and < 20%")]
t <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 40 and < 50%")]
u <- landcover2001@data@attributes[[1]]$ID[which(landcover2001@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 80 and < 90%")]

rangemin <- c(g,f,a,d,h,p,q,s,t)
rangemax <- c(g,f,c,e,o,q,r,t,u)

labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 11, NA), c(18, 22, NA),c(25, 59, NA), c(83,100,NA), c(110,120,NA), c(18, 21, NA), c(129,450,NA))
rcmat <- as.matrix(rc)
rcmat

ras_2001 <- reclassify(landcover2001, rcmat, right=NA)

```

```{r}
#landcover 2008
landcover2008@data@attributes[[1]]$ID
landcover2008@data@attributes[[1]]$CLASSNAMES
length(landcover2008@data@attributes[[1]]$ID)

a <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")] 
b <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")] 
d <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]   
f <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
g <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
h <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Orchard")]
i <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop-Close Grown Crop")]
j <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
k <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Close Grown Crop")]
l <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Fallow/Idle Cropland")]
m <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "NASS-Pasture and Hayland")]
n <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Pasture/Hay")]
o <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Cultivated Crops")]
p <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 10 and < 20%")]
q <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 40 and < 50%")]
r <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 90 and <= 100%")]
s <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 10 and < 20%")]
t <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 40 and < 50%")]
u <- landcover2008@data@attributes[[1]]$ID[which(landcover2008@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 80 and < 90%")]

rangemin <- c(g,f,a,d,h,p,q,s,t)
rangemax <- c(g,f,c,e,o,q,r,t,u)


labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 11, NA), c(18, 22, NA),c(25, 59, NA), c(83,100,NA), c(110,120,NA), c(18, 21, NA), c(129,450,NA))
rcmat <- as.matrix(rc)
rcmat


ras_2008 <- reclassify(landcover2008, rcmat,right=NA)

```

```{r}
#landcover 2010
landcover2010@data@attributes[[1]]$ID
landcover2010@data@attributes[[1]]$CLASSNAMES
length(landcover2010@data@attributes[[1]]$ID)

a <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")] 
b <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")] 
d <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]   
f <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed - Low Intensity")]
g <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
h <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
i <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
j <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 10 and < 20%")]
k <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 40 and < 50%")]
l <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 90 and <= 100%")]
m <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 10 and < 20%")]
n <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 40 and < 50%")]
o <- landcover2010@data@attributes[[1]]$ID[which(landcover2010@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 80 and < 90%")] 

rangemin <- c(g,f,a,d,i,j,k,m,n)
rangemax <- c(h,g,c,e,i,k,l,n,o)

labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 11, NA), c(18, 21, NA), c(25, 63, NA), c(65, 100, NA),c(110, 120, NA),c(129,450,NA),c(31,31,NA))
rcmat <- as.matrix(rc)
rcmat

ras_2010 <- reclassify(landcover2010, rcmat,right=NA)


```


```{r}
#landcover 2012
landcover2012@data@attributes[[1]]$ID
landcover2012@data@attributes[[1]]$CLASSNAMES
length(landcover2012@data@attributes[[1]]$ID)

a <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")] 
b <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")] 
d <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]   
f <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed - Low Intensity")]
g <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
h <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
i <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
j <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "NASS-Close Grown Crop")]
k <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "NASS-Wheat")]
l <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 10 and < 20%")]
m <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 40 and < 50%")]
n <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 90 and <= 100%")]
o <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 10 and < 20%")]
p <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 40 and < 50%")]
q <- landcover2012@data@attributes[[1]]$ID[which(landcover2012@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 80 and < 90%")]  

rangemin <- c(g,f,a,d,i,l,m,o,p)
rangemax <- c(h,g,c,e,k,m,n,p,q)

labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 11, NA),c(25, 63, NA), c(18, 21, NA), c(69,100,NA), c(110,120,NA), c(129,450,NA))
rcmat <- as.matrix(rc)
rcmat

ras_2012 <- reclassify(landcover2012, rcmat,right=NA)

```


```{r}
#landcover 2014
landcover2014@data@attributes[[1]]$ID
landcover2014@data@attributes[[1]]$CLASSNAMES
length(landcover2014@data@attributes[[1]]$ID)

a <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")] 
b <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")] 
d <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]   
f <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed - Low Intensity")]
g <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
h <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
i <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop-Close Grown Crop")]
j <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
k <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "NASS-Close Grown Crop")]
l <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "NASS-Wheat")]
m <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 10 and < 20%")]
n <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 40 and < 50%")]
o <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Tree Cover >= 90 and <= 100%")]
p <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 10 and < 20%")]
q <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 40 and < 50%")]
r <- landcover2014@data@attributes[[1]]$ID[which(landcover2014@data@attributes[[1]]$CLASSNAMES == "Herb Cover >= 80 and < 90%")]  

rangemin <- c(g,f,a,d,i,m,n,p,q)
rangemax <- c(h,g,c,e,l,n,o,q,r)

labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 11, NA), c(18,21,NA), c(25,62,NA), c(69,100,NA), c(109,120,NA),c(129,450,NA))
rcmat <- as.matrix(rc)
rcmat

ras_2014 <- reclassify(landcover2014, rcmat, right=NA)

```

```{r}
#landcover 2016
landcover2016@data@attributes[[1]]$ID
landcover2016@data@attributes[[1]]$CLASSNAMES
length(landcover2016@data@attributes[[1]]$ID)

a <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Deciduous Forest")]
b <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Evergreen Forest")]
c <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Mixed Forest")]
d <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Herbaceous")]
e <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed-Upland Shrubland")]
f <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed - Low Intensity")]
g <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed - Medium Intensity")]
h <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Developed - High Intensity")]
i <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop-Close Grown Crop")]
j <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "NASS-Row Crop")]
k <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "NASS-Close Grown Crop")]
l <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "NASS-Wheat")]
m <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Tree Cover = 10%")]
n <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Tree Cover = 50%")]
o <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Tree Cover = 93%")]
# p <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Shrub Cover = 10%")]
# q <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Shrub Cover = 50%")]
# r <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Shrub Cover = 57%")]
s <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Herb Cover = 10%")]
t <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Herb Cover = 50%")]
u <- landcover2016@data@attributes[[1]]$ID[which(landcover2016@data@attributes[[1]]$CLASSNAMES == "Herb Cover = 91%")]

rangemin <- c(g,f,a,d,i,m,n,s,t)
rangemax <- c(h,g,d,e,l,n,o,t,u)

labels2<- c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA")

rc <- data.frame(cbind(rangemin, rangemax,1:9))
rc <- rbind(rc,c(0, 13, NA), c(18,21,NA), c(25,62,NA),c(69,109,NA), c(193,309,NA), c(392,450,NA))
rcmat <- as.matrix(rc)
rcmat

ras_2016 <- reclassify(landcover2016, rcmat,right=NA)

```


Visualizing the Rasters

```{r}
pal1<- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2016), na.color = "transparent")
pal2 <- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2014), na.color = "transparent")
pal3 <- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2012), na.color = "transparent")
pal4 <- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2010), na.color = "transparent")
pal5 <- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2008), na.color = "transparent")
pal6 <- leaflet::colorFactor(palette = "Spectral", domain = values(ras_2001), na.color = "transparent")


leaflet() %>%
  addRasterImage(ras_2016, pal1, opacity = 1, project = FALSE, group = "Land Cover 2016") %>%
  addRasterImage(ras_2014, pal2, opacity = 1, project = FALSE, group = "Land Cover 2014") %>%
  addRasterImage(ras_2012, pal3, opacity = 1, project = FALSE, group = "Land Cover 2012") %>%
  addRasterImage(ras_2010, pal4, opacity = 1, project = FALSE, group = "Land Cover 2010") %>%
  addRasterImage(ras_2008, pal5, opacity = 1, project = FALSE, group = "Land Cover 2008") %>%
  addRasterImage(ras_2001, pal6, opacity = 1, project = FALSE, group = "Land Cover 2001") %>%
  addLegend(pal = pal4, opacity = 1, values = values(ras_2010), labFormat = function(type, cuts, p) {paste0(labels2)}, group = "Land Cover  Legend", title = "Land Cover Legend") %>%
  addLayersControl(
     overlayGroups = c("Land Cover 2001", "Land Cover 2008", "Land Cover 2010", "Land Cover 2012", "Land Cover 2014", "Land Cover 2016"),
     options = layersControlOptions(collapsed = TRUE))

```

This chunk saves all of the derived dataframes and shapefiles so that the Shiny App can be uploaded to ShinyApps.io or otherwise run independently with the derived data calls and without having to run the .Rmd files.

```{r}
library(raster)

st_write(final, dsn="data/derived/final.shp", delete_dsn= TRUE)
fwrite(as.list(names(final)), file="data/derived/final_colnames.txt")

st_write(Wetlands3, dsn="data/derived/Wetlands3.shp", delete_dsn= TRUE)
fwrite(parks, file="data/derived/parks.txt")


raster::writeRaster(ras_2001, filename=file.path("data/derived/ras_2001.gri"),overwrite=TRUE)
raster::writeRaster(ras_2008, filename=file.path("data/derived/ras_2008.gri"),overwrite=TRUE)
raster::writeRaster(ras_2010, filename=file.path("data/derived/ras_2010.gri"),overwrite=TRUE)
raster::writeRaster(ras_2012, filename=file.path("data/derived/ras_2012.gri"),overwrite=TRUE)
raster::writeRaster(ras_2014, filename=file.path("data/derived/ras_2014.gri"),overwrite=TRUE)
raster::writeRaster(ras_2016, filename=file.path("data/derived/ras_2016.gri"),overwrite=TRUE)

rm(list=ls())
```
