==> 03_Analysis_10_questions.Rmd <==
---
title: "10_Questions"
author: "Bouzaher, Julia"
date: "3/14/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Load in packages

```{r}
library(dplyr)
library(ggplot2)
library(sp)
library(data.table)
library(readxl)
library(sf)
library(raster)
library(leaflet)
library(ggthemes)

```

# Load Data

```{r}
final <- read_sf("data/derived/final.shp")

final_names <- fread("data/derived/final_colnames.txt")

names(final) <- names(final_names)

# names(final) <- gsub("_"," ", names(final))

# names(final)[3] <- "Total Population"

# vars <- names(final)[4:9]

# paste0(vars," Population")

# names(final)[4:9] <- paste0(vars," Population")

final <- final %>%
  # mutate(`Land Area` = ALAND) %>%
  # mutate(`Water Area` = AWATER) %>%
  # rename(`Median Rent` = MedianRent) %>%
  mutate(tract = sub("*, Erie.*", "", tract))
  

ras_2001 <- raster("data/derived/ras_2001.gri")
ras_2008 <- raster("data/derived/ras_2008.gri")
ras_2010 <- raster("data/derived/ras_2010.gri")
ras_2012 <- raster("data/derived/ras_2012.gri")
ras_2014 <- raster("data/derived/ras_2014.gri")
ras_2016 <- raster("data/derived/ras_2016.gri")

park_count <- read_excel("data/derived/park_count.xlsx")
```

Basic Land Cover Per Capita

```{r}
land <- final %>%
  dplyr::select(year, tract, ALAND, Total_Pop, Black) %>%
  unique()

land <- cbind(land, park_count$`Number of County Parks`)

land <- land %>%
  rename(park_count = park_count..Number.of.County.Parks.) 

land <- land %>%
  mutate(park_by_land = (ALAND/park_count)*100) %>%
  mutate(parks_per_capita = 1000*(park_count/as.numeric(Total_Pop))) %>%
  mutate(parks_per_capita_b = 1000*(park_count/as.numeric(Black))) %>%
  filter(tract != "Census Tract 9901")

```

#Figures
```{r}
#over time for total pop
plot1 <- ggplot(data=land, aes(x=as.numeric(year), y=parks_per_capita, fill=tract)) +
  geom_bar(stat="identity", position = position_dodge()) +
  labs(title = "Parks per capita by tract and year", x = "Year", y = "Parks Per Capita (x1000)")

print(plot1 + ggthemes::theme_clean() + scale_x_continuous(breaks=unique(land$year)))
print(plot1 + ggthemes::theme_hc() + scale_x_continuous(breaks=unique(land$year)))
#over time for total pop
plot2 <- ggplot(data=land, aes(x=year, y=parks_per_capita_b, fill=tract)) +
  geom_bar(stat="identity", position = position_dodge()) +
  labs(title = "Parks per capita (Black Individuals only)",  x = "Year", y = "Parks Per Capita (x1000)")

print(plot2 + ggthemes::theme_clean() + scale_x_continuous(breaks=unique(land$year)))
class(land$park_count)
class(land$Total_Pop)

```



Raster tables

```{r}

a <- as.data.frame(freq(ras_2001))
b <- left_join(a, as.data.frame(freq(ras_2008)), by = "value")
c <- left_join(b, as.data.frame(freq(ras_2010)), by = "value")
d <- left_join(c, as.data.frame(freq(ras_2012)), by = "value")
e <- left_join(d, as.data.frame(freq(ras_2014)), by = "value")
f <- left_join(e, as.data.frame(freq(ras_2016)), by = "value")

f <- cbind(f,c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA"))

names(f) <- c("Land Cover ID", "2001 Count", "2008 Count", "2010 Count", "2012 Count", "2014 Count", "2016 Count", "Land Cover Description")
f[,2]

f <- rbind(f, c("Totals", sum(f[,2]), sum(f[,3]), sum(f[,4]), sum(f[,4]), sum(f[,5]), sum(f[,6]), sum(f[,7]), "Totals"))

f[11,8] <- "Totals"

raster_table <- f %>%
  dplyr::select(`Land Cover ID`, `Land Cover Description`, `2001 Count`,  `2008 Count`, `2010 Count`, `2012 Count`, `2014 Count`, `2016 Count`)

counts <- raster_table[,c(3:8)]

counts <- as.data.frame(counts) 

counts <- counts %>%
  mutate_if(is.character, as.numeric)

counts <- (counts/1755948)*100

counts2 <- round(counts, digits =2)

final_table <- as.data.frame(cbind(raster_table$`Land Cover ID`, raster_table$`Land Cover Description`, counts2))

names(final_table) <- c("Land Cover ID","Land Cover Description", "2001 Percentage", "2008 Percentage", "2010 Percentage", "2012 Percentage", "2014 Percentage", "2016 Percentage")

print(final_table)

write.table(final_table, file = "landcoverpercentages.txt", sep = "\t",
            row.names = TRUE, col.names = NA)

```

Point 6: Correlation Matrix
```{r}

table <- data.frame() %>%
  mutate(col_1 = ras_2016@data@attributes[1])

```

Testing how to crop rasters (In order to calculate Raster statistics)

```{r}
land2 <- land %>% 
  filter(tract == "Census Tract 406")

land2 <- st_transform(land2, crs(ras_2016))

tryit <- raster::mask(ras_2016, land2)

leaflet(land) %>%
  addProviderTiles(provider = providers$OpenStreetMap) %>%
  addRasterImage(tryit) %>%
  addPolylines() 
```


Iterative Version

```{r}
years <- c(2001,2008,2010,2012,2014,2016)
tracts <- unique(final$tract)[1:18]

Land_Cover_Series = data.frame()

for (i in years){
  for(j in tracts){
     land2 <- land %>% filter(tract == j)
     land2 <- st_transform(land2, crs(ras_2016))
     clipped <- raster::mask(ras_2016, land2)
     ras <- assign(paste0("ras", i, "_",j), clipped)
     Land_Cover_Series <- as.data.frame(rbind((cbind(freq(ras), year=i, tract = j,Land_Cover=c("Developed - Medium/High Intensity","Developed - Low/Medium Intensity","Developed - Forested", "Developed - herbaceous or shrubland", "Cropland","Tree Cover 0-50%", "Tree Cover 50-100%","Herb Cover 0-50%", "Herb Cover 50-100%", "NA"))))) %>% mutate(count = as.numeric(count))
  }
}

```

#Merging and Analyzing the Land Cover Series
#Correlation Matrices

```{r}

class(Land_Cover_Series$year) <- class(land$year)
Final_Series <- full_join(Land_Cover_Series, land)

library(stats)

Final_Series$Total_Pop <- as.numeric(Final_Series$Total_Pop)

Final_Series <- Final_Series %>%
  group_by(year,tract) %>%
  mutate(year_tract_totals = sum(count)) %>%
  mutate(lc_percentages = round((count/year_tract_totals)*100,2)) %>%
  mutate(black_p = (Black/Total_Pop)*100)

Final_Series <- Final_Series %>%
  dplyr::select(year,tract,park_count,Land_Cover,Black,black_p,parks_per_capita, parks_per_capita_b,lc_percentages, Total_Pop)
```

```{r}

library(Hmisc)
library(stats)

# test <- Final_Series %>%
  # filter(Land_Cover == "Tree Cover 50-100%") %>%
  # filter(Land_Cover == "Tree Cover 0-50%")

row <- as.vector(Final_Series$black_p)
col <- as.vector(Final_Series$lc_percentages)

cor(row,col)

Final_Series <- as.data.frame(Final_Series)
  
test <- Final_Series %>%
  as.data.frame() %>%
  dplyr::select(black_p, parks_per_capita, park_count,Black, Total_Pop,lc_percentages)

#graphing examination
ggplot(test, aes(x = black_p, y = lc_percentages )) +
  geom_jitter() +
  geom_smooth() +
  geom_jitter(aes(x = black_p, y = parks_per_capita)) +
  geom_smooth()


png(filename="Std_PNG.png", 
    units="in", 
    width=5, 
    height=4, 
    pointsize=12, 
    res=300)

#correlation tests
library(corrplot)
M <- cor(test)

corrplot(M, type = "lower", outline = T, addgrid.col = "darkgray", bg="azure", title = "Correlation")
dev.off()

```




Demographics questions

```{r}
`%notin%` <- Negate(`%in%`)

graphdata <- final %>%
  filter(tract %notin% c("Census Tract 418", "Census Tract 9901", "Census Tract 417", "Census Tract 405", "Census Tract 406", "Census Tract 401", "Census Tract 402","Census Tract 404","Census Tract 403", "Census Tract 416"))

plot2 <- ggplot(graphdata, aes(x=year,y=Black, col=tract)) +
  geom_jitter(col="black") + 
  geom_smooth(se = FALSE) +
  labs(title = "Black population changes over time by census tract")

plot2 + ggthemes::theme_clean()

graphdata2 <- final %>%
  filter(tract %in% c("Census Tract 407", "Census Tract 408", "Census Tract 410", "Census Tract 411", "Census Tract 412", "Census Tract 413"))

pchanges <- graphdata2 %>%
  arrange(year, tract) %>%
  group_by(tract) %>%
  mutate(changerent = (MedianRent/lead(MedianRent) - 1) * 100,
         changeb = (Black/lead(Black) - 1) * 100,
         changeincome = (`Mean household income`/lead(`Mean household income`) - 1) * 100,
         )

plot4 <- ggplot(pchanges, aes(x=year, y=changerent, fill = tract)) +
  
  geom_bar(stat = "identity", position = "stack",aes(fill = "red",y=changeb))+
  geom_bar(stat="identity", position = "stack", aes(fill="blue")) +
  geom_bar(stat = "identity", position = "stack",aes(fill = "green",y=changeincome))+
  labs(x = "Year", y = "Percent change", title = "Percent changes in topics of interest by tract", fill = "Topic") +
  facet_wrap(~tract)

plot4 +
  ggthemes::theme_clean() +
  theme(legend.position = "bottom") +
  scale_fill_discrete(labels = c("Black Population","Mean Household Income","Median Rent"))

plot3 <- ggplot(graphdata2, aes(x=year,y=Black, col=tract)) +
  geom_jitter(col="black") + 
  geom_smooth(se = FALSE) +
  geom_smooth(aes(y=Total_Pop),se = FALSE) +
 # geom_smooth(aes(y = graphdata2$MedianRent)) +
  #geom_smooth(aes(y = graphdata2$`Mean household income`)) +
  labs(title = "Black population changes over time by census tract") +
  facet_wrap(~tract)

plot3 + ggthemes::theme_clean()
  

```

Final Race Table

```{r}
#(MedianRent/lead(MedianRent) - 1) * 100
plot4 <- ggplot(graphdata2, aes(x=year, y=MedianRent)) +
  geom_bar(stat="identity", fill = "orange",position="stack",aes(y=White/5))+
  geom_bar(stat="identity", fill = "blue",position="stack",aes(y=Black/5))+
  
  geom_bar(stat="identity", fill = "red",position="stack",aes(y=Other))+
  geom_bar(stat="identity", fill = "green",position="stack",aes(y=Asian))+
  
  geom_bar(stat="identity", fill = "pink",position="stack",aes(y=Native_Hawaiian_or_Pacific_Islander))+
  geom_bar(stat="identity", fill = "purple",position="stack",aes(y=Native_American))+
  geom_smooth(col = "black", se = FALSE) +
  
  # geom_smooth(aes(y = graphdata2$MedianRent)) +
  #geom_smooth(aes(y = graphdata2$`Mean household income`)) + 
  labs(title = "Demographic changes and Median Rent Changes") +
  facet_wrap(~tract) 

plot4 + ggthemes::theme_clean()

```
Saving Final Table

```{r}

names(final) <- gsub("_"," ", names(final))

names(final)[3] <- "Total Population"

vars <- names(final)[4:9]

names(final)[4:9] <- paste0(vars," Population")

final_t <- final %>%
  mutate(`Land Area` = ALAND) %>%
  mutate(`Water Area` = AWATER) %>%
  rename(`Median Rent` = MedianRent) %>%
  mutate(tract = sub("*, Erie.*", "", tract)) %>%
  dplyr::select(-c(ALAND,AWATER))

st_write(final_t, dsn="data/derived/final_table.shp", delete_dsn= TRUE)
fwrite(as.list(names(final_t)), file="data/derived/final_table_colnames.txt")

```
